// src/typing/enums.ts
export enum NodeEnv {
  production = 'production',
  test = 'test',
  development = 'development',
}

export enum LogLevel {
  error = 'error',
  warn = 'warn',
  info = 'info',
  http = 'http',
  verbose = 'verbose',
  debug = 'debug',
  silly = 'silly',
}

export enum LogFile {
  combined = 'combined',
  error = 'error',
}

export enum Product {
  productOne = 'Product One',
  productTwo = 'Product Two',
  productThree = 'Product Three',
}

export enum Status {
  notStarted = 'Not Started',
  processing = 'Processing',
  sent = 'Sent',
}

export enum DataSet {
  customers = 'customers',
  orders = 'orders',
}